Fix APIParams serialisation issue
